Split a bunch of Foolscap stuff off into its own module
